feat: AddedStepRVAdpater interface, ViewHolder 별도로 분리 #226
